NetBSD: Fix the field type of `statvfs.f_spare`
Link: https://github.com/NetBSD/src/blob/6017cb90fd7d83ed6e45d93129b12d525978c2fa/sys/sys/statvfs.h#L94

+++ src/unix/bsd/netbsdlike/netbsd/mod.rs
@@ -882,7 +882,7 @@ s_no_extra_traits! {
         pub f_namemax: c_ulong,
         pub f_owner: crate::uid_t,
 
-        pub f_spare: [u32; 4],
+        pub f_spare: [u64; 4],
 
         pub f_fstypename: [c_char; 32],
         pub f_mntonname: [c_char; 1024],
